move todos sort function to todosFunction, clean up comments
